RestaurantsPage test 추가 및 구현
App과 RestaurantPage 분리
